Refactoring/add script to generate test data
